fix: SocketRelay code-review batch — audit evidence, chat guard, tag notice, parse dedup
Low-risk cleanup from the code-review sweep. No schema, route, or contract
change; audit metadata now matches the already-declared contracts.

- Audit evidence fields (#1885, #1886, #1883): request.create now records
  roleCheck/payloadValidationCheck; fulfillment.claim records
  claimabilityCheck/ownerSeparationCheck; service-credits.send records
  recipientUserId/amountPositiveCheck. Reaching each audit call proves the
  corresponding checks passed, so the values are recorded as passed.
- Chat-credential guard (#1882): the web Direct Line renders the Stream chat
  only when streamToken and streamUserId are both present, dropping the
  `as string` casts, so a partial response falls back to the empty pane
  instead of handing undefined to the Stream SDK.
- Tag-trim notice (#1890): the post form shows "Tag trimmed to 64 characters."
  when it shortens an over-long tag instead of changing the input silently.
- Parse-input dedup (#1891): parseTags / parsePriceAmount / parseRequestInput
  were duplicated in the create and update routes; they now live once in
  lib/socket-relay/parse-input.ts.

Inventory change log and the SR-3 test-script note updated to match.
