+++ ctf/docs/developer/ctf-plugin-feature-inventories/ctf-socket-relay-feature-inventory.md
@@ -175,6 +175,8 @@ alongside the legacy `category`) and fulfillment outcomes for dev validation.
 
 ## 9) Change Log
 
+- 2026-07-26: **Code-review batch — audit evidence, chat-credential guard, tag-trim notice, parse-input dedup.** Low-risk cleanup from the code-review sweep. (a) The three member-mutation audit rows now carry the evidence fields their audit contracts ask for: `socket-relay.request.create` adds `roleCheck`/`payloadValidationCheck`, `socket-relay.fulfillment.claim` adds `claimabilityCheck`/`ownerSeparationCheck`, and `socket-relay.service-credits.send` adds `recipientUserId`/`amountPositiveCheck` (#1885, #1886, #1883) — reaching each audit call proves those checks passed, so the values are recorded as passed. (b) The web Direct Line now renders the Stream chat only when `streamToken` and `streamUserId` are both present (dropping the `as string` casts), so a partial credentials response falls back to the empty pane instead of handing `undefined` to the Stream SDK (#1882). (c) The post form's tag editor shows a brief "Tag trimmed to 64 characters." notice when it shortens an over-long tag, instead of silently changing the input (#1890). (d) `parseTags` / `parsePriceAmount` / `parseRequestInput` were duplicated in the create and update routes; they now live once in `lib/socket-relay/parse-input.ts` so the two paths cannot drift (#1891). No schema, route, or contract change; audit metadata now matches the already-declared contract.
+
 - 2026-07-26: **Members-only visibility (no public board) + open-scoped feed with "Load more".** Two owner-directed fixes. (1) **Visibility drift.** The list and single-request routes disagreed: the feed (`GET /requests`) showed every request to any signed-in member, but the detail route (`GET /requests/:id`) still ran a v2 `is_public` gate and returned 403 to a non-owner on a non-public request the feed had just listed. Per the owner ruling — v3 has no public post; only signed-in members can view, including a shared deep link — the detail route's `is_public` visibility check is removed (any signed-in member may view any request), and the anonymous public board is deleted: `GET /api/socket-relay/public`, `GET /api/socket-relay/public/:id`, the `SocketRelayPublicRequest` DTO, and the `listPublicRequests` / `getPublicRequestById` / `mapPublicRequestRow` functions are gone. The `is_public` column is retained but inert (no read path consults it; it was already un-toggled since 2026-06-12). (2) **Feed pagination.** `listRequests` gained an optional `statuses` filter and `GET /requests` an optional `?status=` param; the web feed now asks for `status=open` (claimable) so resolved/claimed posts don't crowd open ones off the first 20, with a **"Load more"** button that appends the next page (de-duped by id). The "N open" badge reads the server-side open total. "Mine" already sources `my-requests` (all statuses), so a member's own resolved/claimed posts stay reachable. Schema: none (`is_public` kept). Contracts/test-script/inventory updated.
 
 - 2026-07-26: **Full code-review sweep of SocketRelay — correctness & reliability batch.** From a comprehensive review (the automated sweep timed out before the contracts/last files). Fixes: (a) the "Mine" feed filter now sources the owner-scoped `my-requests` list (fetched at pageSize 100) instead of the 20-item global feed, so a member's own older posts no longer vanish from "Mine" (they were being hidden once 20 newer board-wide posts existed, blocking Edit/Re-post); (b) the resolve route now returns proper status codes — a helper trying to resolve gets 403 (`actor_not_requester`), an already-resolved fulfillment gets 409 (`fulfillment_not_active`), and a bad `outcome` gets 400 (`invalid_outcome`) — instead of a 503 that also fired a false server-error alert; (c) `repostRequest` is now transactional with a `FOR UPDATE` lock and a `status <> 'claimed'` predicate (closing a claim-vs-repost race the earlier unlocked guard left open) and writes a `request_reposted` lifecycle event (closes #1881); (d) the claim-time Stream channel call is now best-effort — a Stream outage no longer fails an already-committed claim, drops its audit row, or loses the "someone offered to help" notification (the chat route self-heals the channel); (e) `updateRequest` guards against a concurrent delete (clean `request_not_found` instead of a 500); (f) the feed empty state now says "No matches" for a search/filter with no results instead of falsely claiming the board is empty; (g) the `?fulfillment=` deep-link param is stripped after handoff so a refresh doesn't re-force the Direct Line tab; (h) added the missing unique index backing `socket_relay_messages`' `ON CONFLICT` idempotency (the send route would otherwise throw 42P10). Verified false: #1367 (`updateRequest` already emits its `request_updated` event). Schema: one unique index (no columns/data). Contract-doc drift (undeclared profile/repost/admin-delete commands, resolve audit event, deletion-contract fictional tables, a few `dataAccess` mismatches) is tracked as a separate follow-up.
