+++ ctf/packages/web/app/api/socket-relay/requests/route.ts
@@ -14,7 +14,8 @@ import {
   listRequests,
   validateRequestInput,
 } from 'lib/socket-relay/repository';
-import type { SocketRelayRequestInput, SocketRelayRequestStatus } from 'lib/socket-relay/types';
+import type { SocketRelayRequestStatus } from 'lib/socket-relay/types';
+import { parseRequestInput } from 'lib/socket-relay/parse-input';
 import { reportError } from 'lib/observability/report';
 
 const REQUEST_STATUSES: SocketRelayRequestStatus[] = ['open', 'claimed', 'closed', 'cancelled'];
@@ -30,49 +31,6 @@ function parseStatusFilter(raw: string | null): SocketRelayRequestStatus[] | und
       (REQUEST_STATUSES as string[]).includes(value),
     );
   return wanted.length > 0 ? wanted : undefined;
-}
-
-// Only a real number or a non-empty numeric string becomes an amount; booleans, arrays, objects, and
-// `null`/`undefined` never coerce to a price (so e.g. `true` is not read as 1).
-function parsePriceAmount(value: unknown): number | null {
-  if (typeof value === 'number') {
-    return Number.isFinite(value) && value > 0 ? value : null;
-  }
-  if (typeof value === 'string' && value.trim().length > 0) {
-    const parsed = Number(value.trim());
-    return Number.isFinite(parsed) && parsed > 0 ? parsed : null;
-  }
-  return null;
-}
-
-// Older clients send a single `category` string; newer ones send a `tags` array (1-3).
-function parseTags(body: Record<string, unknown>): string[] {
-  if (Array.isArray(body.tags)) {
-    return body.tags.filter((tag): tag is string => typeof tag === 'string');
-  }
-  return typeof body.category === 'string' && body.category.trim() ? [body.category] : [];
-}
-
-function parseRequestInput(body: Record<string, unknown>): SocketRelayRequestInput {
-  // Value type (issue #420): a non-empty currency code names how the request is settled; an absent/blank
-  // code means none was chosen. Amount is only kept as a positive finite number; anything else is null
-  // (so amount-less types like Free/Barter carry no amount).
-  const priceCurrency =
-    typeof body.priceCurrency === 'string' && body.priceCurrency.trim().length > 0
-      ? body.priceCurrency.trim()
-      : null;
-  const priceAmount = parsePriceAmount(body.priceAmount);
-  return {
-    title: typeof body.title === 'string' ? body.title : '',
-    details: typeof body.details === 'string' ? body.details : '',
-    tags: parseTags(body),
-    city: typeof body.city === 'string' ? body.city : null,
-    state: typeof body.state === 'string' ? body.state : null,
-    country: typeof body.country === 'string' ? body.country : null,
-    isPublic: typeof body.isPublic === 'boolean' ? body.isPublic : false,
-    priceCurrency,
-    priceAmount,
-  };
 }
 
 export async function GET(request: Request) {
@@ -146,6 +104,9 @@ export async function POST(request: Request) {
       reason: 'ok',
       targetType: 'request',
       targetId: item.id,
+      // Evidence fields the audit contract asks for. Reaching this call proves the read-access gate and
+      // the payload validation above both passed.
+      metadata: { roleCheck: 'pass', payloadValidationCheck: 'pass' },
     });
     return NextResponse.json({ ok: true, item }, { status: 201 });
   } catch (error) {
